refactor: extract composables and type net-worth-chart
