+++ frontend/common/src/premium/index.ts
@@ -111,18 +111,22 @@ export interface SettingsApi {
   };
 }
 
-interface GradientArea {
-  color: {
-    colorStops: {
-      color: string;
-      offset: number;
-    }[];
-    type: string;
-    x: number;
-    x2: number;
-    y: number;
-    y2: number;
-  };
+interface ColorStop {
+  color: string;
+  offset: number;
+}
+
+interface GradientColor {
+  colorStops: ColorStop[];
+  type: 'linear';
+  x: number;
+  x2: number;
+  y: number;
+  y2: number;
+}
+
+export interface GradientArea {
+  color: GradientColor;
 }
 
 export interface NewGraphApi {
@@ -141,14 +145,21 @@ export interface GraphApi {
   gridColor: ComputedRef<string>;
 }
 
-type GetGraphApi = (canvasId: string) => GraphApi;
-
-type GetNewGraphApi = () => NewGraphApi;
+interface GetGraphApi {
+  /**
+   * Initializes and returns an instance of GraphApi for a specified canvas element.
+   *
+   * @param {string} canvasId - The ID of the canvas element where the graph will be rendered.
+   * @returns {GraphApi} An instance of the GraphApi that provides methods to interact with the graph.
+   * @deprecated
+   */
+  (canvasId: string): GraphApi;
+  (): NewGraphApi;
+}
 
 export interface PremiumApi {
   readonly date: DateUtilities;
   readonly data: DataUtilities;
   readonly settings: SettingsApi;
   readonly graphs: GetGraphApi;
-  readonly newGraphs: GetNewGraphApi;
 }
